// backend/src/common/enums/role.enum.ts
export enum VaiTroEnum {
  ADMIN = 'ADMIN',
  DIRECTOR = 'DIRECTOR',
  TEAM_LEAD = 'TEAM_LEAD',
  SALES_REP = 'SALES_REP',
  MARKETING = 'MARKETING',
  CUST_SUCCESS = 'CUST_SUCCESS',
  ACCOUNTANT = 'ACCOUNTANT',
}

export enum PhamViDuLieuEnum {
  MY_DATA = 'MY_DATA',       // Dữ liệu của tôi
  TEAM_DATA = 'TEAM_DATA',   // Dữ liệu của nhóm tôi
  ALL_DATA = 'ALL_DATA',     // Tất cả dữ liệu
}